requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9; python_version < '3.13'
psycopg[binary]==3.2.3; python_version >= '3.13'
requests==2.31.0
supabase==2.0.3
tzdata>=2024.1
